+++ lib/src/mock_document_reference.dart
@@ -5,7 +5,7 @@ import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
 
 import 'converter.dart';
 import 'mock_document_reference_platform.dart';
-import 'cloud_firestore_mocks_base.dart';
+import 'fake_cloud_firestore_instance.dart';
 import 'mock_collection_reference.dart';
 import 'mock_document_snapshot.dart';
 import 'mock_field_value_platform.dart';
@@ -20,7 +20,7 @@ class MockDocumentReference<T extends Object?> implements DocumentReference<T> {
   final Map<String, dynamic> docsData;
   final Map<String, dynamic> rootParent;
   final Map<String, dynamic> snapshotStreamControllerRoot;
-  final MockFirestoreInstance _firestore;
+  final FakeFirebaseFirestore _firestore;
   final Converter<T>? _converter;
 
   /// Path from the root to this document. For example "users/USER0004/friends/FRIEND001"